# src/main/java/com/signify/reviews/dto/request/ReviewFilterDto.java
package com.signify.reviews.dto.request;

import lombok.*;
import org.springframework.format.annotation.DateTimeFormat;

import javax.validation.constraints.Max;
import javax.validation.constraints.Min;

@Setter
@Getter
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class ReviewFilterDto {

    @Min(1)
    @Max(5)
    private Integer rating;
    private String store;

    @DateTimeFormat(pattern = "yyyy-MM-dd")
    private String startDate;

    @DateTimeFormat(pattern = "yyyy-MM-dd")
    private String endDate;
//    private Integer page;
//    private Integer pageSize;
}