# src/main/java/com/signify/reviews/dto/response/TotalRating.java
package com.signify.reviews.dto.response;

import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Setter
@Getter
@NoArgsConstructor
@AllArgsConstructor
public class TotalRating {
    private Integer rating;
    private Integer total;
}